Added canvas
Added Hello World on the canvas.

The text is supposed to show in the middle of the canvas, therefore canvas.width() and canvas.height() are used. We take half of the value to catch the middle. Since the text start at this point, in the width we have to account for the length of the text to center it. The -50 takes care of that. This is omitted for the height.